+++ include/primitives/mesh.h
@@ -11,12 +11,12 @@ class Mesh
 {
 public:
 	std::vector<Vertex> m_vertices;
-	std::vector<GLuint> m_indices;
+	std::vector<GLushort> m_indices;
 	std::vector<Texture> m_textures;
 
 	VAO m_vao;
 
-	Mesh(std::vector<Vertex> vertices, std::vector<GLuint> indices, std::vector<Texture> textures);
+	Mesh(std::vector<Vertex> vertices, std::vector<GLushort> indices, std::vector<Texture> textures);
 
 	void Bind();
